requirements.txt
coverage==5.0.4
Django==3.1.3
django-rest-auth==0.9.5
django-extensions==3.0.9
djangorestframework==3.12.2
factory-boy==2.12.0
flake8==3.8.4
Pillow==8.0.1
psycopg2==2.8.4
python-memcached==1.59
PyYAML==5.3.1
